OCPBUGS-57419: Do not log "new CO status" unnecessarily
Currently, we log this even when the status is not actually new.
